Coin Flip Gui
Initial release of the coin flipping program.  This small program takes a single user input (non-zero, positive integer) and "flips a coin" an input value amount of times.  The program then displays 5 values, how many times the coin landed on "heads", how many times it landed on "tails", the percentage chance of landing on either based on the values and how long it took to run through those calculations.